+++ src/skyperious/conf.py
@@ -24,7 +24,7 @@ import appdirs
 
 """Program title, version number and version date."""
 Title = "Skyperious"
-Version = "5.4.dev12"
+Version = "5.4.dev13"
 VersionDate = "20.09.2022"
 
 if getattr(sys, "frozen", False):