+++ src/skyperious/lib/util.py
@@ -138,6 +138,37 @@ def format_exc(e):
     return result
 
 
+def format_sql_name(name):
+    """Quotes name for use in an SQL expression, if name needs quoting."""
+    SQLITE_KEYWORDS = [
+        "ABORT", "ACTION", "ADD", "AFTER", "ALL", "ALTER", "ALWAYS", "ANALYZE",
+        "AND", "AS", "ASC", "ATTACH", "AUTOINCREMENT", "BEFORE", "BEGIN",
+        "BETWEEN", "BY", "CASCADE", "CASE", "CAST", "CHECK", "COLLATE", "COLUMN",
+        "COMMIT", "CONFLICT", "CONSTRAINT", "CREATE", "CROSS", "CURRENT", 
+        "CURRENT_DATE", "CURRENT_TIME", "CURRENT_TIMESTAMP", "DATABASE", "DEFAULT",
+        "DEFERRABLE", "DEFERRED", "DELETE", "DESC", "DETACH", "DISTINCT", "DO",
+        "DROP",  "EACH", "ELSE", "END", "ESCAPE", "EXCEPT", "EXCLUDE", "EXCLUSIVE",
+        "EXISTS",  "EXPLAIN", "FAIL", "FILTER", "FIRST", "FOLLOWING", "FOR",
+        "FOREIGN", "FROM",  "FULL", "GENERATED", "GLOB", "GROUP", "GROUPS",
+        "HAVING", "IF", "IGNORE",  "IMMEDIATE", "IN", "INDEX", "INDEXED",
+        "INITIALLY", "INNER", "INSERT",  "INSTEAD", "INTERSECT", "INTO", "IS",
+        "ISNULL", "JOIN", "KEY", "LAST", "LEFT", "LIKE", "LIMIT", "MATCH",
+        "MATERIALIZED", "NATURAL", "NO", "NOT", "NOTHING", "NOTNULL", "NULL",
+        "NULLS", "OF", "OFFSET", "ON", "OR", "ORDER", "OTHERS", "OUTER", "OVER",
+        "PARTITION", "PLAN", "PRAGMA", "PRECEDING", "PRIMARY", "QUERY", "RAISE",
+        "RANGE", "RECURSIVE", "REFERENCES", "REGEXP", "REINDEX", "RELEASE",
+        "RENAME", "REPLACE", "RESTRICT", "RETURNING", "RIGHT", "ROLLBACK", "ROW",
+        "ROWS", "SAVEPOINT", "SELECT", "SET", "TABLE", "TEMP", "TEMPORARY", "THEN",
+        "TIES", "TO", "TRANSACTION", "TRIGGER", "UNBOUNDED", "UNION", "UNIQUE",
+        "UPDATE", "USING", "VACUUM", "VALUES", "VIEW", "VIRTUAL", "WHEN", "WHERE",
+        "WINDOW", "WITH", "WITHOUT",
+    ]
+    result = to_unicode(name)
+    if result.upper() in SQLITE_KEYWORDS or re.search(r"(^[\W\d])|(?=\W)", result, re.U):
+        result = u'"%s"' % result.replace('"', '""')
+    return result
+
+
 def format_sql_value(value):
     """Formats value to be suitable for use in an SQL expression."""
     UNPRINTABLES = "".join(set(six.unichr(i) for i in range(128)).difference(string.printable))