+++ CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
 - add configuration flag to auto-open exported files
 - add 1:1 and group chat message counts to contact details
 - add context menu popup to contact list right-click
+- add support for exporting SQL query results as INSERT statements
 - export contact list by current applied filter
 - order file types in save dialog by label
 - skip trying to download old shared content guaranteed to be unavailable